+++ example/lib/main.dart
@@ -4,8 +4,10 @@ import 'dart:ui';
 import 'package:example/lang_view.dart';
 import 'package:example/my_flutter_app_icons.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:easy_localization/easy_localization.dart';
+import 'package:easy_localization_loader/easy_localization_loader.dart';
+
+import 'generated/locale_keys.g.dart';
 
 void main() {
   runApp(EasyLocalization(
@@ -16,17 +18,23 @@ void main() {
       Locale('de', 'DE'),
       Locale('ru', 'RU')
     ],
-    path: 'resources/langs',
+    path: 'resources/langs/langs.csv', //'resources/langs',
     // fallbackLocale: Locale('en', 'US'),
     // saveLocale: false,
     // useOnlyLangCode: true,
+    // preloaderColor: Colors.black,
+
     // optional assetLoader default used is RootBundleAssetLoader which uses flutter's assetloader
+    // install easy_localization_loader for enable custom loaders
     // assetLoader: RootBundleAssetLoader()
-    // assetLoader: NetworkAssetLoader()
-    // assetLoader: TestsAssetLoader()
+    // assetLoader: HttpAssetLoader()
     // assetLoader: FileAssetLoader()
-    // assetLoader: StringAssetLoader()
-    // preloaderColor: Colors.red,
+    assetLoader: CsvAssetLoader()
+    // assetLoader: YamlAssetLoader() //multiple files
+    // assetLoader: YamlSingleAssetLoader() //single file
+    // assetLoader: XmlAssetLoader() //multiple files
+    // assetLoader: XmlSingleAssetLoader() //single file
+    // assetLoader: CodegenLoader()
   ));
 }
 
@@ -38,11 +46,7 @@ class MyApp extends StatelessWidget {
     log('title'.tr().toString(), name: '${this} # locale');
     return MaterialApp(
       title: 'title'.tr(),
-      localizationsDelegates: [
-        GlobalMaterialLocalizations.delegate,
-        GlobalWidgetsLocalizations.delegate,
-        EasyLocalization.of(context).delegate,
-      ],
+      localizationsDelegates:EasyLocalization.of(context).delegates,
       supportedLocales: EasyLocalization.of(context).supportedLocales,
       locale: EasyLocalization.of(context).locale,
       theme: ThemeData(
@@ -82,7 +86,7 @@ class _MyHomePageState extends State<MyHomePage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('title').tr(context: context),
+        title: Text(LocaleKeys.title).tr(context: context),
         //Text(AppLocalizations.of(context).tr('title')),
         actions: <Widget>[
           FlatButton(
@@ -105,14 +109,14 @@ class _MyHomePageState extends State<MyHomePage> {
               flex: 1,
             ),
             Text(
-              'switch.with_arg',
+              LocaleKeys.gender_with_arg,
               style: TextStyle(
                   color: Colors.grey.shade600,
                   fontSize: 19,
                   fontWeight: FontWeight.bold),
             ).tr(args: ['aissat'], gender: _gender ? 'female' : 'male'),
             Text(
-              tr('switch', gender: _gender ? 'female' : 'male'),
+              tr(LocaleKeys.gender, gender: _gender ? 'female' : 'male'),
               style: TextStyle(
                   color: Colors.grey.shade600,
                   fontSize: 15,
@@ -129,19 +133,20 @@ class _MyHomePageState extends State<MyHomePage> {
             Spacer(
               flex: 1,
             ),
-            Text('msg').tr(args: ['aissat', 'Flutter']),
-            Text('clicked').plural(counter),
+            Text(LocaleKeys.msg).tr(args: ['aissat', 'Flutter']),
+            Text(LocaleKeys.msg_named).tr(namedArgs: {'lang': 'Dart'}, args: ['Easy localization']),
+            Text(LocaleKeys.clicked).plural(counter),
             FlatButton(
               onPressed: () {
                 incrementCounter();
               },
-              child: Text('clickMe').tr(),
+              child: Text(LocaleKeys.clickMe).tr(),
             ),
             SizedBox(
               height: 15,
             ),
             Text(
-                plural('amount', counter,
+                plural(LocaleKeys.amount, counter,
                     format: NumberFormat.currency(
                         locale: Intl.defaultLocale, symbol: '€')),
                 style: TextStyle(
@@ -155,7 +160,7 @@ class _MyHomePageState extends State<MyHomePage> {
               onPressed: () {
                 EasyLocalization.of(context).deleteSaveLocale();
               },
-              child: Text('reset_locale').tr(),
+              child: Text(LocaleKeys.reset_locale).tr(),
             ),
             Spacer(
               flex: 1,