+++ lib/src/asset_loader.dart
@@ -1,11 +1,13 @@
 import 'dart:convert';
 import 'dart:developer';
+import 'dart:ui';
 
+import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/services.dart';
 
 abstract class AssetLoader {
   const AssetLoader();
-  Future<Map<String, dynamic>> load(String localePath);
+  Future<Map<String, dynamic>> load(String path, Locale locale);
 }
 
 //
@@ -15,9 +17,14 @@ abstract class AssetLoader {
 class RootBundleAssetLoader extends AssetLoader {
   const RootBundleAssetLoader();
 
+  String getLocalePath(String basePath, Locale locale) {
+    return '$basePath/${localeToString(locale, separator: "-")}.json';
+  }
+
   @override
-  Future<Map<String, dynamic>> load(String localePath) async {
-    log('easy localization: Load asset from $localePath');
+  Future<Map<String, dynamic>> load(String path, Locale locale) async {
+    var localePath = getLocalePath(path, locale);
+    log('easy localization: Load asset from $path');
     return json.decode(await rootBundle.loadString(localePath));
   }
 }